services/dev-auth-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
